requirements.txt
#Core LangChain and Vector Store
langchain
langchain-community
chromadb

#Sentence Transformers for Text Embeddings
sentence-transformers

#Hugging Face Transformers and PyTorch for the Language Model
transformers
torch

#Document Processing Libraries
PyPDF2
python-docx
pytesseract
Pillow
googletrans==4.0.0rc1
pdf2image

#Performance Libraries for Transformers
accelerate
bitsandbytes
safetensors